onboarding/onboarding_requirements.txt
# Onboarding System Requirements - From conda environment "onboarding"
# Generated from actual conda environment on 2025-09-26

# ===== CORE FRAMEWORK & WORKFLOW =====
langgraph==0.6.6
langchain==0.3.27
langchain-openai==0.3.32
langchain-core==0.3.74
langchain-community==0.3.27
langchain-postgres==0.0.15
langchain-text-splitters==0.3.9
langgraph-checkpoint==2.1.1
langgraph-checkpoint-sqlite==2.0.11
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.3
langsmith==0.4.17

# ===== DJANGO INTEGRATION =====
django==5.2
django-taggit==6.1.0
djangorestframework==3.16.0
asgiref==3.9.1
sqlparse==0.5.3

# ===== WEB FRAMEWORK =====
fastapi==0.116.1
starlette==0.47.3
uvicorn==0.35.0
python-multipart==0.0.20
h11==0.16.0

# ===== DATABASE & CACHE =====
asyncpg==0.30.0
redis==6.4.0
sqlalchemy==2.0.43
psycopg==3.2.9
psycopg2-binary==2.9.10
psycopg-pool==3.2.6
aiosqlite==0.21.0
pgvector==0.3.6
sqlite-vec==0.1.6
greenlet==3.2.4

# ===== LLM & AI =====
openai==1.101.0
tiktoken==0.11.0
ollama==0.5.3

# ===== RESEARCH TOOLS =====
wikipedia==1.4.0
wikipedia-api==0.8.1
arxiv==2.2.0
requests==2.32.5
requests-file==2.1.0
requests-toolbelt==1.0.0
beautifulsoup4==4.13.5
readability-lxml==0.8.4.1
lxml==6.0.1
lxml-html-clean==0.4.2
feedfinder2==0.0.4
feedparser==6.0.11

# ===== CONTENT PROCESSING =====
python-readability==0.1.3
html2text==2025.4.15
newspaper3k==0.2.8
tldextract==5.3.0
cssselect==1.3.0
chardet==5.2.0
sgmllib3k==1.0.0
charset-normalizer==3.4.3
jieba3k==0.35.1
tinysegmenter==0.3
nltk==3.9.1

# ===== SEARCH APIs =====
serpapi==0.1.5
google-search-results==2.4.2

# ===== ASYNC PROCESSING =====
aiohttp==3.12.15
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
aiodns==3.5.0
nest-asyncio==1.6.0
anyio==4.10.0
multidict==6.6.4
frozenlist==1.7.0
yarl==1.20.1
propcache==0.3.2
pycares==4.10.0

# ===== DATA VALIDATION & PROCESSING =====
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic-core==2.33.2
marshmallow==3.26.1
dataclasses-json==0.6.7

# ===== JUPYTER & DEVELOPMENT =====
jupyter==1.1.1
ipykernel==6.30.1
ipython==9.4.0
ipywidgets==8.1.7
notebook==7.4.5
jupyterlab==4.4.6
notebook-shim==0.2.4
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.8.1
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter-server==2.17.0
jupyter-server-terminals==0.5.3
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.15
jupyterlab-pygments==0.3.0
widgetsnbextension==4.0.14
ipython-pygments-lexers==1.1.1

# ===== VISUALIZATION =====
matplotlib==3.10.5
matplotlib-inline==0.1.7
plotly==6.3.0
python-graphviz==0.21
pygraphviz==1.14
fonttools==4.59.1
contourpy==1.3.3
cycler==0.12.1
kiwisolver==1.4.9
pillow==11.3.0

# ===== DATA ANALYSIS =====
pandas==2.3.2
numpy==2.3.2
narwhals==2.2.0

# ===== NOTEBOOK & DOCUMENT PROCESSING =====
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
mistune==3.1.3
pandocfilters==1.5.1
bleach==6.2.0
defusedxml==0.7.1
fastjsonschema==2.21.2
tinycss2==1.4.0
webencodings==0.5.1

# ===== UTILITIES =====
pyyaml==6.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
click==8.2.1
jinja2==3.1.6
markupsafe==3.0.2
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.1
mypy-extensions==1.1.0
types-python-dateutil==2.9.0.20250822
arrow==1.3.0
pytz==2025.2
tzdata==2025.2
packaging==25.0
six==1.17.0
filelock==3.19.1
platformdirs==4.3.8

# ===== NETWORK & HTTP =====
httpx==0.28.1
httpcore==1.0.9
httpx-sse==0.4.1
certifi==2025.8.3
urllib3==2.5.0
idna==3.10
sniffio==1.3.1
websocket-client==1.8.0

# ===== JSON & SERIALIZATION =====
orjson==3.11.2
ormsgpack==1.10.0
json5==0.12.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
python-json-logger==3.3.0

# ===== TERMINAL & SYSTEM =====
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
terminado==0.18.1
send2trash==1.8.3
psutil==7.0.0
pexpect==4.9.0
ptyprocess==0.7.0
wcwidth==0.2.13
prompt-toolkit==3.0.51

# ===== DEVELOPMENT & DEBUGGING =====
debugpy==1.8.16
executing==2.2.0
pure-eval==0.2.3
stack-data==0.6.3
asttokens==3.0.0
decorator==5.2.1
parso==0.8.5
jedi==0.19.2
pygments==2.19.2

# ===== PARSING & VALIDATION =====
lark==1.2.2
pyparsing==3.2.3
soupsieve==2.7
attrs==25.3.0
annotated-types==0.7.0
referencing==0.36.2
rpds-py==0.27.0
overrides==7.7.0
tenacity==9.1.2

# ===== WEB VALIDATION =====
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
uri-template==1.3.0
fqdn==1.5.1
webcolors==24.11.1
isoduration==20.11.0

# ===== COMPRESSION & ENCODING =====
brotli==1.1.0
zstandard==0.24.0
xxhash==3.5.0

# ===== SECURITY & AUTHENTICATION =====
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
cffi==1.17.1
pycparser==2.22

# ===== ASYNC & CACHING =====
async-lru==2.0.5
comm==0.2.3
pyzmq==27.0.2

# ===== MONITORING & LOGGING =====
prometheus-client==0.22.1
distro==1.9.0
babel==2.17.0

# ===== TEXT PROCESSING =====
regex==2025.7.34
joblib==1.5.1

# ===== JAVASCRIPT BRIDGE =====
pythonmonkey-fork==1.2.0
pythonmonkey-node-modules==0.1.2

# ===== MACHINE LEARNING EXTENSIONS =====
jiter==0.10.0